Creates seed data for Heroes. Adds 'role' to hero

+++ backend/db/migrate/20210218174814_create_heros.rb
@@ -2,7 +2,7 @@ class CreateHeros < ActiveRecord::Migration[6.0]
   def change
     create_table :heros do |t|
       t.string :name
-      t.integer :winrate
+      t.float :winrate
       t.integer :teamID
 
       t.timestamps
