+++ backend/db/migrate/20210218174911_create_matches.rb
@@ -2,7 +2,7 @@ class CreateMatches < ActiveRecord::Migration[6.0]
   def change
     create_table :matches do |t|
       t.string :winner
-      t.integer :winnerProbability
+      t.float :winnerProbability
 
       t.timestamps
     end
